+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/gateway
 
+## 4.0.0-beta.27
+
+### Patch Changes
+
+- 70a9aae: feat (provider/gateway): add disallowPromptTraining gateway provider option
+
 ## 4.0.0-beta.26
 
 ### Patch Changes